# docs/06-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The authenticated user"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      },
      "patch": {
        "operationId": "AuthController_updateMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The updated profile"
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/admin/supervisor/me": {
      "get": {
        "operationId": "SupervisorController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The signed-in staff member and the grant held"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Caller holds no active platform_staff grant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/supervisor/cases": {
      "get": {
        "operationId": "SupervisorController_list",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "AI operations cases, projected from audit_logs"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Caller holds no active platform_staff grant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/supervisor/cases/{id}": {
      "get": {
        "operationId": "SupervisorController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One case with live domain state and timeline"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Caller holds no active platform_staff grant"
          },
          "404": {
            "description": "No AI operations case with this id"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/supervisor/cases/{id}/resolve": {
      "post": {
        "operationId": "SupervisorController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Case resolved; one append-only audit row written"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Caller holds no active platform_staff grant"
          },
          "404": {
            "description": "No AI operations case with this id"
          },
          "409": {
            "description": "Case was already resolved"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "BANHAO API",
    "description": "BANHAO | บ้านเฮา — Phase 1 Food Delivery API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}